Исправление ошибок. Модуляция
