+++ js/data.js
@@ -1,3 +1,5 @@
+import {generateRandom} from './util.js';
+
 const DESCRIPTIONS = [
   'Это я в Туриции была, сейчас уже дома',
   'Всем хорошего настроения в этот замечательный день',
@@ -25,17 +27,41 @@ const NAMES = [
   'Артём'
 ];
 
-function generateRandom(a, b) {
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
+function generateComment(number) {
+  const message = Array(generateRandom(1, 2));
+  for (let i = 0; i < message - 1; i++) {
+    message[i] = COMMENTS[generateRandom(0, COMMENTS.length - 1)];
+  }
 
-  return Math.floor(result);
+  return {
+    id: number,
+    avatar: `img/avatar-${generateRandom(1, 6)}.svg`,
+    message: message,
+    name: NAMES[generateRandom(0, NAMES.length - 1)]
+  };
 }
 
-export {generateRandom, DESCRIPTIONS, COMMENTS, NAMES};
+function generateDescription(identifier){
+  const comments = Array(3);
+  for (let i = 0; i < 3; i++) {
+    comments[i] = generateComment(i + 1);
+  }
 
-function checkLength(verifiedString, maxLength) {
-  return (verifiedString.length >= maxLength);
+  return {
+    id: identifier,
+    url: 'photos/identifier.jpg',
+    descriptions: DESCRIPTIONS[generateRandom(0, DESCRIPTIONS.length - 1)],
+    likes: generateRandom(15, 200),
+    comments: comments
+  };
 }
-checkLength('Привет!', 10);
+
+function generateDescriptions(){
+  const descriptions = Array(25);
+  for (let i = 0; i < 25; i++) {
+    descriptions[i] = generateDescription(i + 1);
+  }
+  return (descriptions);
+}
+
+export {generateDescriptions};
